Move team chat to child component of session
